# Reorganize arg bundles in the runtime.