--- a/modules/internal/LocaleModelHelpFlat.chpl
+++ b/modules/internal/LocaleModelHelpFlat.chpl
@@ -67,7 +67,7 @@ module LocaleModelHelpFlat {
       // the compiler calls this version in some cases.
       chpl_ftable_call(fn, args);
     } else {
-      var tls = chpl_task_getChapelData();
+      var tls = chpl_task_getInfoChapel();
       chpl_task_data_setup(chpl_comm_on_bundle_task_bundle(args), tls);
       chpl_comm_execute_on(node, chpl_sublocFromLocaleID(loc),
                            fn, args, args_size);
@@ -90,7 +90,7 @@ module LocaleModelHelpFlat {
       // don't call the runtime fast execute_on function if we can stay local
       chpl_ftable_call(fn, args);
     } else {
-      var tls = chpl_task_getChapelData();
+      var tls = chpl_task_getInfoChapel();
       chpl_task_data_setup(chpl_comm_on_bundle_task_bundle(args), tls);
       chpl_comm_execute_on_fast(node, chpl_sublocFromLocaleID(loc),
                                 fn, args, args_size);
@@ -112,7 +112,7 @@ module LocaleModelHelpFlat {
     // non-blocking "on" in order to serialize the execute_ons.
     //
     const node = chpl_nodeFromLocaleID(loc);
-    var tls = chpl_task_getChapelData();
+    var tls = chpl_task_getInfoChapel();
     var isSerial = chpl_task_data_getSerial(tls);
     if (node == chpl_nodeID) {
       // don't call the runtime nb execute_on function if we can stay local