--- a/runtime/include/chpl-arg-bundle.h
+++ b/runtime/include/chpl-arg-bundle.h
@@ -58,4 +58,3 @@ size_t chpl_argBundleSizeofHdr(void* arg) {
 #endif // LAUNCHER
 
 #endif
-