--- a/runtime/include/chpl-comm.h
+++ b/runtime/include/chpl-comm.h
@@ -246,7 +246,7 @@ void chpl_comm_rollcall(void);
 //   the memory did indeed come from chpl_mem_regMemAlloc(), this frees
 //   it and returns true.  Otherwise it does nothing and returns false.
 //   Given some memory address to be freed it is therefore safe, though
-//   perhaps not performance-optimal, to first try to free it here, and 
+//   perhaps not performance-optimal, to first try to free it here, and
 //   only free it elsewhere if this function returns false.
 //
 #ifndef CHPL_COMM_IMPL_REG_MEM_HEAP_INFO
@@ -375,7 +375,7 @@ void chpl_comm_broadcast_private(int id, size_t size);
 // cannot be immediately satisfied, while it waits chpl_comm_barrier()
 // must call chpl_task_yield() in order not to monopolize the execution
 // resources and prevent making progress. This barrier must be available
-// for use in module code, so it cannot be tied up in the runtime 
+// for use in module code, so it cannot be tied up in the runtime
 //
 void chpl_comm_barrier(const char *msg);
 
@@ -433,8 +433,8 @@ void chpl_comm_get(void *addr, c_nodeid_t node, void* raddr,
 //            and strides.
 // When comm=gasnet, this function ends up calling gasnet_puts_bulk().
 //   More info in: http://www.escholarship.org/uc/item/5hg5r5fs?display=all
-//   Proposal for Extending the UPC Memory Copy Library Functions and Supporting 
-//   Extensions to GASNet, Version 2.0. Author: Dan Bonachea 
+//   Proposal for Extending the UPC Memory Copy Library Functions and Supporting
+//   Extensions to GASNet, Version 2.0. Author: Dan Bonachea
 //
 void chpl_comm_put_strd(void* dstaddr, size_t* dststrides, c_nodeid_t dstnode,
                         void* srcaddr, size_t* srcstrides, size_t* count,
@@ -521,4 +521,3 @@ void chpl_wait_for_shutdown(void);
 #include "chpl-comm-warning-macros.h"
 
 #endif
-