--- a/runtime/include/comm/ugni/chpl-comm-task-decls.h
+++ b/runtime/include/comm/ugni/chpl-comm-task-decls.h
@@ -42,9 +42,15 @@ typedef struct {
 // Comm layer private area within executeOn argument bundles
 // (bundle.comm)
 typedef struct {
+  chpl_bool fast;
   chpl_fn_int_t fid;
-  int caller;
+  c_nodeid_t caller;
+  c_sublocid_t subloc;
+  size_t size;
   void* rf_done; // where to indicate completion on caller
+#ifdef CHPL_COMM_DEBUG
+  uint_least64_t seq;
+#endif
 } chpl_comm_bundleData_t;
 
 // The type of the communication handle.