--- a/runtime/include/tasks/fifo/chpl-tasks-impl-fns.h
+++ b/runtime/include/tasks/fifo/chpl-tasks-impl-fns.h
@@ -27,8 +27,6 @@
 
 #include <stdint.h>
 
-#include "chpl-tasks-prvdata.h"
-
 #ifdef __cplusplus
 extern "C" {
 #endif