--- a/runtime/include/tasks/fifo/chpl-tasks-impl.h
+++ b/runtime/include/tasks/fifo/chpl-tasks-impl.h
@@ -67,7 +67,7 @@ typedef struct {
   chpl_fn_int_t requested_fid;
   chpl_fn_p requested_fn;
   chpl_taskID_t id;
-  chpl_task_ChapelData_t state;
+  chpl_task_infoChapel_t infoChapel;
 } chpl_task_bundle_t;