--- a/runtime/src/chpl-cache.c
+++ b/runtime/src/chpl-cache.c
@@ -2964,4 +2964,3 @@ void chpl_cache_set_enabled(int enabled)
 
 #endif
 // end ifdef HAS_CHPL_CACHE_FNS
-