The bundled arguments for outlined tasking and on-stmt body functions
consist of a runtime-defined header struct, followed by the arguments
themselves.  There are distinct header struct types for tasking and
on-stmt arg bundles.

On-stmt, or "comm", bundle header structs have two uses.  They support
comm layers using Active Messages to invoke on-stmt function bodies on
remote nodes, of course, but they also support internal Active Messages
for support activities such as coordinating program shutdown and doing
GETs and PUTs that cannot use comm layer RMA features.  In comm=gasnet
the opcode, if you will, of an Active Message is delivered out-of-band
from the AM's contents.  But in comm=ugni and comm=ofi the underlying
communication library either doesn't have such support or we don't use
it, so the opcode has to be embedded in the AM message itself.  This
being so, it has to be in the same place in all AM messages, so that the
target-side AM handler can pick it out of the received message and
figure out how to proceed from there.

It used to be the case that the header struct for an on-stmt arg bundle
began with a member that was the header struct for a tasking arg bundle
and then followed that with some additional information.  Thus all arg
bundles, of either kind, effectively began with the tasking header.
This allowed tasking layers to find their information in either kind of
bundle, which is necessary because the tasking layer actually runs the
body functions of both tasking and on-stmts.  (All comm layers initiate
on-stmt body functions using the tasking layer on the target node.)  The
downside of this was that comm arg bundles for internal AM opcodes, as
opposed to actual on-stmts, were excessively large because they had a
task bundle header struct on the front that they didn't need.

Here, reorganize the arg bundle headers.  Put a small member encoding
the bundle header type first in both kinds of bundles, with room in the
encoding for a reasonable number of comm layer-specific internal AM
opcodes.  This allows identifying the type of the bundle just by looking
at this initial member.  Then, rearrange the contents of the comm header
so that the comm info comes first and the task header comes second.  Now
task arg bundles still consist of a tasking header struct followed by
the bundled arguments, and on-stmt arg bundles consist of a comm header
struct followed by a tasking bundle (tasking header + args).

While a fair amount of code changed here, there are really only a few
functional changes.

The tasking header is no longer necessarily the first thing in any
bundle.  Thus the code cannot simply type-pun an arg bundle pointer to
reach the tasking information.  Instead it has to identify the bundle
header type and then get at the tasking info using the correct one.

Ugni and ofi comm layer-specific internal AM messages are now smaller,
since they can consist of just the initial layer-specific header type
code plus only those members needed for the internal AM activity in
question.  This is not a big change; to a very rough approximation these
messages are now in the 40-80 byte range instead of the 150-200 byte
range, but it's still a marginal improvement.

While here, I also took the opportunity to greatly simplify how both
comm=ugni and ofi build and handle on-stmt arg bundles, in particular
the divergence between on-stmt arg bundles that fit in a maximum size AM
request and those that do not.

Also while here, I removed the special wrapper for the "main" task in
tasks=qthreads and replaced its use with the regular task wrapper, did
some runtime include file restructuring, and fixed up a few whitespace
flaws.

This resolves Cray/chapel-private#926 and resolves #12350 and
resolves #12364.