@ronawho 2020-05-12
>> When the machine opens up in a few hours I'd like to do a coforall+on scalability run to sanity check performance.

> Yeah, sounds good, and I'll wait for results on that. I spot-checked performance and saw neither degradation nor much improvement except to ofi large coforall+on Remote Task Spawn Time (which has been terrible up until now), but it would be good to have a full run to make sure I didn't miss something.

Just to be clear -- I was only planning on running the coforall+on microbenchmarks at 512 nodes of an XC. I did not plan on doing a full perf run, I'm happy to leave that to nightly.