@gbtitus 2020-05-12
> Just to be clear -- I was only planning on running the coforall+on microbenchmarks at 512 nodes of an XC.

Ah, I misunderstood.  I think I'll go ahead and merge this, then.  I'm good with letting nightlies do the full perf run also.