@gbtitus 2020-05-12
Oops!  I'm going to need one more commit.  I've just discovered some flaws that only appear in `DEBUG=1` builds.