@ronawho 2020-05-12
coforall+on perf looks good. Minor improvements for empty and average bundle sizes

---

Time to do 100K coforall+ons on 512 nodes of an XC (ugni):

```chpl
var argBundle: bundleSize*int(8);
for 1..100_000 do
  coforall loc in Locales with (in argBundle) do on loc { }
```

| bundleSize  | before | after  |
| ----------: | -----: | -----: |
|    0        |  32.5s |  32.2s |
|  512        |  38.4s |  36.3s |
| 4096        | 123.5s | 123.5s |

Where at least as of late 2018 512 byte bundles were the average user bundle size.